sos_project/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
pip==24.2
psycopg2-binary==2.9.10
setuptools==65.5.0
sqlparse==0.5.1
